feat(core): add ICounterTrackable interface and strongly-typed AddCou…

+++ RuleEngine/src/RuleEngine.Domain/CrewManagement/Entities/CounterValue.cs
@@ -3,8 +3,8 @@
     public class CounterValue
     {
         public string CounterTypeSystemName { get; set; }
-        public int CounterValue_ { get; set; }
-        public CounterValue(string counterTypeSystemName, int counterValue)
+        public decimal CounterValue_ { get; set; }
+        public CounterValue(string counterTypeSystemName, decimal counterValue)
         {
             CounterTypeSystemName = counterTypeSystemName;
             CounterValue_ = counterValue;
